bugfix(weapon): Fix unreliability of historic bonus weapons (#1727)
This change fixes historic bonus weapons not triggering reliably. The issue is that whenever a historic bonus weapon is triggered (e.g. a firestorm) it resets the historic bonus count for all weapons using the same template. This means that if a firestorm is triggered while another firestorm is half way to triggering, that other firestorm's progress will be reset.

In addition, historic damage instances now more accurately expire based on the weapon template's HistoricBonusTime rather than the global HistoricDamageLimit. Historic damage involved in triggering a historic weapon cannot be used in subsequent attempts.